requirements.txt
groq==0.33.0
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
